% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_data_eta}
\alias{generate_data_eta}
\title{generate_data}
\usage{
generate_data_eta(alpha, eta)
}
\arguments{
\item{alpha}{Intercept parameters}

\item{eta}{Linear predictor for each row}
}
\value{
Simulated data y of length length(eta)
}
\description{
Generate data where each participant is a single row.
}
