% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_data_x}
\alias{generate_data_x}
\title{generate_data_x}
\usage{
generate_data_x(x, Xdes, alpha, beta)
}
\arguments{
\item{x}{Participant intervention indicators}

\item{Xdes}{Design matrix}

\item{alpha}{Intercept parameters}

\item{beta}{Coefficient parameters}
}
\value{
Simulated dataset
}
\description{
Generate data where each participant assignment is x.
}
