% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inv_ord_logit}
\alias{inv_ord_logit}
\title{inv_ord_logit}
\usage{
inv_ord_logit(a)
}
\arguments{
\item{x}{A vector of non-decreasing real numbers}
}
\value{
Inverse ordered logit of x
}
\description{
Calculate inverse ordered logit
}
