% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asthmasims.R
\name{sim_asthma_trial}
\alias{sim_asthma_trial}
\title{sim_asthma_trial}
\usage{
sim_asthma_trial(
  mod,
  n_seq = 200,
  p_assign = rep(1/3, 3),
  alpha = stats::qlogis(cumsum(rep(1/101, 101)))[-101],
  eta = c(0, 0, 0),
  eff_eps = 0.975,
  ...
)
}
\arguments{
\item{mod}{A list (model, model data)}

\item{n_seq}{Sequence of interim analysis sample sizes}

\item{p_assign}{Assignment probabilities to treatment arms}

\item{alpha}{True intercept parameter}

\item{eta}{True eta parameter}

\item{eff_eps}{Effectiveness threshold}

\item{...}{Other arguments to cmdstanr::sample, e.g. adapt_delta, chains, etc.}
}
\value{
A list of trial related data.tables
}
\description{
Simulate asthma trial
}
\details{
Simulates asthma trial.
}
