% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_data_X}
\alias{generate_data_X}
\title{generate_data_X}
\usage{
generate_data_X(X, alpha, beta)
}
\arguments{
\item{X}{Design matrix}

\item{alpha}{Intercept parameters}

\item{beta}{Design matrix parameters}
}
\value{
Simulated data y of length nrow(X)
}
\description{
Generate data where each participant is a single row.
}
